Split `Menu` and `SubMenu` styles.
