+++ src/menu/styles/menu.m.css
@@ -1,11 +1,6 @@
 @import '../../common/styles/variables.css';
 
 .root {
-	box-sizing: border-box;
-	position: relative;
-}
-
-.menu {
 	background: var(--menu-bg);
 	box-shadow: 0 0 2px 1px rgba(0,0,0,0.3);
 	font-family: sans-serif;
@@ -13,7 +8,7 @@
 	padding: 5px 0;
 }
 
-.menu:focus {
+.root:focus {
 	outline: none;
 }
 
@@ -45,8 +40,6 @@
 	color: var(--menu-selected-color);
 }
 
-.nestedMenuRoot {}
-
 .nestedMenu {
 	box-sizing: border-box;
 	padding: 0 15px;
@@ -55,67 +48,4 @@
 .horizontal {
 	display: flex;
 	flex-direction: row;
-}
-
-.subMenu {}
-.inline {}
-
-.dropDown,
-.popup {
-	position: absolute;
-}
-
-.dropDown.hidden,
-.popup.hidden {
-	visibility: hidden;
-}
-
-/* Animations */
-@keyframes fadeIn {
-	0% {
-		opacity: 0;
-		visibility: hidden;
-	}
-	100% {
-		opacity: 1;
-		visibility: visible;
-	}
-}
-@keyframes fadeOut {
-	0% {
-		opacity: 1;
-		visibility: visible;
-	}
-	100% {
-		opacity: 0;
-		visibility: hidden;
-	}
-}
-.fade.hidden { animation: fadeOut var(--short-animation-duration) ease-in-out both; }
-.fade.visible { animation: fadeIn var(--short-animation-duration) ease-in-out both; }
-
-.slide {
-	overflow: hidden;
-	transition: height var(--short-animation-duration) ease-in-out;
-	width: 100%;
-}
-.slide.hidden { height: 0; }
-.slide.visible { height: auto; }
-
-/* Positioning */
-.left {
-	left: auto;
-	right: 100%;
-}
-.right {
-	left: 100%;
-	right: auto;
-}
-.bottom {
-	bottom: auto;
-	top: 100%;
-}
-.top {
-	bottom: 100%;
-	top: auto;
 }
