+++ src/menu/styles/menu.m.css.d.ts
@@ -1,23 +1,7 @@
 export const root: string;
-export const menu: string;
 export const menuItem: string;
 export const menuLabel: string;
 export const disabled: string;
 export const selected: string;
-export const nestedMenuRoot: string;
 export const nestedMenu: string;
 export const horizontal: string;
-export const subMenu: string;
-export const inline: string;
-export const dropDown: string;
-export const popup: string;
-export const hidden: string;
-export const fade: string;
-export const fadeOut: string;
-export const visible: string;
-export const fadeIn: string;
-export const slide: string;
-export const left: string;
-export const right: string;
-export const bottom: string;
-export const top: string;
