+++ src/menu/styles/subMenu.m.css
@@ -1,0 +1,69 @@
+@import '../../common/styles/variables.css';
+
+.root {
+	box-sizing: border-box;
+	position: relative;
+}
+
+.subMenu {}
+.inline {}
+
+.dropDown,
+.popup {
+	position: absolute;
+}
+
+.dropDown.hidden,
+.popup.hidden {
+	visibility: hidden;
+}
+
+/* Animations */
+@keyframes fadeIn {
+	0% {
+		opacity: 0;
+		visibility: hidden;
+	}
+	100% {
+		opacity: 1;
+		visibility: visible;
+	}
+}
+@keyframes fadeOut {
+	0% {
+		opacity: 1;
+		visibility: visible;
+	}
+	100% {
+		opacity: 0;
+		visibility: hidden;
+	}
+}
+.fade.hidden { animation: fadeOut var(--short-animation-duration) ease-in-out both; }
+.fade.visible { animation: fadeIn var(--short-animation-duration) ease-in-out both; }
+
+.slide {
+	overflow: hidden;
+	transition: height var(--short-animation-duration) ease-in-out;
+	width: 100%;
+}
+.slide.hidden { height: 0; }
+.slide.visible { height: auto; }
+
+/* Positioning */
+.left {
+	left: auto;
+	right: 100%;
+}
+.right {
+	left: 100%;
+	right: auto;
+}
+.bottom {
+	bottom: auto;
+	top: 100%;
+}
+.top {
+	bottom: 100%;
+	top: auto;
+}
