+++ src/menu/styles/subMenu.m.css.d.ts
@@ -1,0 +1,15 @@
+export const root: string;
+export const subMenu: string;
+export const inline: string;
+export const dropDown: string;
+export const popup: string;
+export const hidden: string;
+export const fade: string;
+export const fadeOut: string;
+export const visible: string;
+export const fadeIn: string;
+export const slide: string;
+export const left: string;
+export const right: string;
+export const bottom: string;
+export const top: string;
